+++ src/tools/x509/impl/sign.cpp
@@ -50,7 +50,7 @@ std::vector<uint8_t> sign(const void* buffer, size_t size, Sha2::Type sha, EVP_P
         EVP_DigestSignUpdate(ctx, buffer, size);
 
         // Compute signature
-        signature.resize(EVP_PKEY_size(pkey));
+        signature.resize(static_cast<size_t>(EVP_PKEY_size(pkey)));
         size_t sig_size = signature.size();
         if (EVP_DigestSignFinal(ctx, &signature[0], &sig_size) == 1)
         {
@@ -92,11 +92,11 @@ std::vector<uint8_t> sign(const std::string& filepath, Sha2::Type sha, EVP_PKEY*
             do
             {
                 file.read(reinterpret_cast<char*>(buffer), sizeof(buffer));
-                EVP_DigestSignUpdate(ctx, buffer, file.gcount());
+                EVP_DigestSignUpdate(ctx, buffer, static_cast<size_t>(file.gcount()));
             } while (file.gcount() == sizeof(buffer));
 
             // Compute signature
-            signature.resize(EVP_PKEY_size(pkey));
+            signature.resize(static_cast<size_t>(EVP_PKEY_size(pkey)));
             size_t sig_size = signature.size();
             if (EVP_DigestSignFinal(ctx, &signature[0], &sig_size) == 1)
             {
@@ -174,7 +174,7 @@ bool verify(const std::vector<uint8_t>& signature, const std::string& filepath, 
             do
             {
                 file.read(reinterpret_cast<char*>(buffer), sizeof(buffer));
-                EVP_DigestVerifyUpdate(ctx, buffer, file.gcount());
+                EVP_DigestVerifyUpdate(ctx, buffer, static_cast<size_t>(file.gcount()));
             } while (file.gcount() == sizeof(buffer));
 
             // Verify signature
